refactor(strategy): migrate Strategy.ID and run IDs from int to int64
Aligns with Postgres BIGINT id columns and the rest of the codebase
(users.ID, portfolio ids), removing int(int64) casts at bridge points
and clearing CodeQL int-conversion warnings on inbound form IDs.

- Strategy.ID, StrategyRun.ID, StrategyRun.StrategyID, RunResult.StrategyID,
  BacktestConfig.StrategyID all become int64
- Repository methods (GetByID, Update, Delete, GetRuns, GetLatestRun,
  GetRunStats) take int64
- Handlers parse path params with strconv.ParseInt(s, 10, 64) directly
- DashboardStrategy.ID and DashboardRun.StrategyID in views updated to match

+++ internal/strategy/backtest.go
@@ -4,7 +4,7 @@ import "time"
 
 // BacktestConfig defines the parameters for running a backtest
 type BacktestConfig struct {
-	StrategyID     int       `json:"strategy_id"`
+	StrategyID     int64     `json:"strategy_id"`
 	StartDate      time.Time `json:"start_date"`
 	EndDate        time.Time `json:"end_date"`
 	RebalanceFreq  string    `json:"rebalance_freq"` // "monthly", "quarterly", "semi-annual", "annual"
